refactor(shorts): 홈페이지 스크롤 캐로셀 구조 리팩토링
• VideoCarousel 컴포넌트를 shared에서 home으로 이동
• ShortformDetail 모달 컴포넌트 제거
• 비디오 클릭 시 동적 라우트(/shortform/[videoId])로 이동하도록 변경
• 캐러셀 loop 옵션을 false로 변경하여 무한 스크롤 비활성화
• 불필요한 state 관리 로직 제거
